LAB_7/thuc-hanh/requirements.txt
dash
Dash
html
dcc
dash_core_components
plotly.express
pandas
firebase_admin
credentials
firestore